blogrecommendation/requirements.txt
python-dotenv
pika
pymongo
nltk
requests
